vision/mae/pytorch/requirements.txt
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
timm==0.5.4
transformers==4.11.3
torchvision==0.14.1+cpu; python_version > '3.6'
torchvision==0.11.0+cpu; python_version <= '3.6'
protobuf==3.19.4
pytest==6.2.4
pyyaml>=5.4.1
horovod>=0.24.0
simplejpeg>=1.6.4
wandb==0.12.1
git+https://github.com/graphcore/examples-utils@latest_stable
numpy==1.23.5; python_version > '3.6'
numpy==1.19.5; python_version <= '3.6'
